Ensure that we show errors and we are logged in.
The old Auth did not update at all times, we changed this API to check the token more often. We also did some login checking, all of this is born due to the issue that we just store the token in localStorage, but we are unsure if it's valid or not.