Login state
Login state have been rewritten a bit, now we will make sure to update all tokens before running the app code, the app code will also redirect to the login if we get an unathued code form the api.

This should make it more stable, and when a token expires it should be seamless to get a new one.

Now we also allow to view pages before the team leader has picked either  KG or KR.